@@ -18,16 +18,33 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
+  meta: ^1.3.0
   flutter:
     sdk: flutter
-
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  flutter_screenutil: ^5.0.0
+  share_extend: ^2.0.0
+ # multi_image_picker: ^4.8.0
+  image_picker: ^0.7.4
+  path_provider: ^2.0.1
+  image_gallery_saver: ^1.6.8
+  permission_handler: ^6.0.1
+  fluttertoast: ^8.0.1-nullsafety.0
+  dio: ^4.0.0-beta6
+  web_socket_channel: ^2.1.0
+  uuid: ^3.0.4
+  device_info: ^2.0.0
+  flutter_native_timezone: ^1.0.10
+  http: ^0.13.0
+  rxdart: ^0.26.0-nullsafety.1
+  shared_preferences: ^2.0.1
+  flutter_local_notifications: ^5.0.0+4
 
 dev_dependencies:
   flutter_test:
